Переписал всё
