--- requirements.txt
@@ -1,14 +1,0 @@
-beautifulsoup4==4.10.0
-bs4==0.0.1
-certifi==2021.10.8
-charset-normalizer==2.0.7
-decorator==5.1.0
-greenlet==1.1.2
-idna==3.3
-requests==2.26.0
-six==1.16.0
-soupsieve==2.2.1
-SQLAlchemy==1.4.25
-urllib3==1.26.7
-validators==0.18.2
-Werkzeug==2.0.1
